+++ client/src/hooks/usePlayer.ts
@@ -2,11 +2,11 @@ import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { GET, GET_TEAM_ID } from '@/app/api/subscribe/route';
 import { GET as GET_TEAM } from '@/app/api/team/route';
 import { gamerInfo, Team } from '@/types';
-import { useToast } from '@/hooks/use-toast';
+import { useToast } from '@/hooks/useToast';
 import { useEffect, useState } from 'react';
 import { supabase } from '@/utils/supabase/client';
 import { TABLES } from '@/constant/db';
-import { useUserId } from '@/utils/hooks/userAuth';
+import { useUserId } from '@/hooks/userAuth';
 
 export function usePlayerList(team: string) {
   const { toast } = useToast();
